refactor: rename cardModal to cardPage component
