+++ react-components/src/pages/Home/components/cardPage/cardPage.scss
